Extract common sorting methods into RichEvent
